pbtest: delete pkg because upstream IDL changes.
This commit deletes the blackbox end-to-end tests associated with
`testing/quick`, because several incompatible upstream changes broke
them, and it is difficult to tease apart where the fault lies for now.
It fixes #9.

The core of this is code is small and well tested.  At the very worst,
integrators can inform us quickly about breakages.